fix: Prevent LeftPanel from changing commit selection
The LeftPanel would override the user's commit selection whenever
its content would get refreshed. The tree is handling `AfterSelect` event
and that causes the user's selection to change to whatever branch or tag
is being selected in the LeftPanel's tree.

Stop this from happening by ignoring the event handling sequence whenever
the tree is reloaded.

Fixes #5144