fix hypothetical race condition
In some circumstances GE may crash with

System.InvalidOperationException: Process was not started by this object,
so requested information cannot be determined.

The most likely scenario is Thread1 can pass the `if (_disposed == 0)`
(line 107) and then Thread2 can come into play and call `Dispose` until
the line 191. After that Thread1 gets resumed and works in the disposed
state.

Closes #5362