Remove `SizeChanged` handler
The current implementation of the `SizeChanged` handler causes a bad
distortion to the UI and renders the listview unusable.

The handler is removed and the column is resized at the time of bindings
items to the list.
This adds a small artifact that if the list was resized after items have
been bound, the column may no longer be of the full width of the list.

Fixes #5437
Fixes #5438