Fix SubmoduleInfoResult having no CurrentSubmoduleName for submodules of submodules
Problem is that when GetSubmodulesLocalPaths builds recursive paths, it doesn't convert the path to a posix path, which is done everywhere else. Later, in GetSuperProjectRepositorySubmodulesStatusAsync, when it compares paths to determine the CurrentSubmoduleName, the paths are not in the same form, and thus fail the comparison.

As a side effect/benefit, the names of sub-submodules now look better in the display as they no longer contain backslashes.