Fix RemoteObjectTree branch folder nodes sometimes becoming bold
This bug was introduced in cd68a511068cd2c0095793d27b2f4c50efb955a6 where setting TreeViewNode.NodeFont = AppSettings.Font was removed as an optimization. Problem is that for performance reasons, nodes are recycled, and it's possible for a node that was a LocalBranchNode for the active branch (thus, with bold font) to become a BranchPathNode. This would leave the BranchPathNode with bold font instead of the regular font.

This change brings back the former change, except this time we only update the node's font if required.