Allow esc key in commit diff to close FormCommit (#5634)
Previously, pressing escape in either file list of the message box would
dismiss FormCommit. However if focus was in the diff view, escape would
not.

This commit fixes this behaviour. It also ensures if text is selected,
the first esc keypress clears the selection, as previously happened.

This event is currently only used in FormCommit but may be useful
elsewhere.