Fix RepoObjectsTree (left panel) remaining blank when ReloadAsync is cancelled (e.g. by spamming refresh)
Problem is that if during ReloadAsync, a System.OperationCanceledException is thrown due to cancellation, we cannot use token anymore in the finally block to switch to main thread. Doing so causes another exception to be thrown, and we never end up executing the code below, which includes "treeMain.EndUpdate()" required to restore rendering of the tree view.