SubmoduleStatusProvider always sets TopProject
Prior to this change, SubmoduleStatusProvider would only set TopProject if current module's SuperProject was not the TopProject. In other words, when the current module was at least 2 submodules deep from the top module. With this change, we always set the TopProject to the top module. This means that SuperProject == TopProject when current module is 1 submodule deep from the top module. This change makes it easier to retrieve the top project when needed at no extra runtime cost.