Throw OperationCanceledException on cts.Next() on disposed cts.
CancellationTokenSequence is written in a thread-safe manner but
it can not be accessed when it is disposed.
Throw OperationCanceledException instead of ObjectDisposedException
so it is possible to access cts from a Task without thinking of cts being disposed or not.

(cherry picked from commit 7e9494993228713394f322da0060e09687b2a436)