[#6147] Select rows in the oder they had originally been selected
So far, the grid attempts to select a commit as soon as it is loaded.
Consequently, rows were being selected in the order in which revisions
are read, not in the order in which they had originally been selected.
By waiting till all rows to be selected have been added to the grid
before attempting to select them, we can select them in the correct
order. With the current implementation, that's the best that can be done
if we don't want to wait till all commits have been loaded before
attempting to select the rows.