Fix #6170 by no longer using local paths
The code used to check whether the file existed in the current directory
before resolving it from the actual Git repository directory. This was
problematic as it led to PathTooLongExceptions for the intermediate
(and usually bogus) result of current dir + file name even if the real
file path wasn't too long at all.

Also unify with the other occurrence of "new FileInfo" in this file for
consistency (which already ignored the current directory). Remove
Path.GetFullPath as FileInfo.FullName does the same.