fix: Throws on `git remote` call outside git repo
It looks like there was a change in wording after git 2.16 that
changed from "fatal: Not a git..." to "fatal: not a git...".

The existing parser expected the warning to be in the lower case, and
thus failed for earlier versions of git.

The fix makes the check case insensitive.

Fixes #6562
Closes #6563
Closes #6565
Closes #6570
Closes #6571
Closes #6574
Closes #6577
Closes #6571