refactor: GitModule.GetSortedRefs
`GitModule.GetSortedRefs` method has been moved into `CommitInfo` control
because that's was its only call site. As such the method added little
value and a significant overhead to the public API surface.

To get the tests working,
* `MockExecutable` was moved to `CommonTestUtils`
projects to it can now be accessed across all tests; and
* `LoadSortedRefsAsync()` method was refactored and `ToDictionary` method
was inlined into `GetSortedRefs()` to facilitate testing.

Also while building tests I discovered the following:
* `CommitInfo` would call `RefreshSortedRefs()` when the control was
passed an instance of `UICommandsSource`, which was pretty much
at the instantiation time, way before a display of a commit information
was requested.
* A second redundant call to `RefreshSortedRefs()` was made when a repository
was closed and a user navigated to the dashboard.
* `UpdateCommitMessage()` method could get called with a `null` revision,
and that would cause a NRE.