Add a `PseudoSetting` to display controls that are not settings
i.e. that don't save a value in the settings files.
It is used to easily display a control that is not a setting (linklabel, text,...)

PS: Solve problem linklabels truncated horizontally of vertically if the good size is not set
(because the size no more have to be set)