Fix unit test on `TryGetExactPathName()` (for AppVeyor VS2019 image)
The test that fails is partly due to a refactoring made in 1246b50 where I raise a little the bar with the expectations.

Indeed, before that, when the files existing, the unit test was verifying... nothing!

I made a change to verify that the paths cases was the ones of the real paths. I was thinking that the windows paths would have always the same cases for system files. It seems not to be the case :( as the AppVeyor VS2019 image has a different one...

I can propose is to find a in-between between testing nothing (like before the refactoring) and having the expected casing (after the refactoring).

I.e:

    We start with the paths without the good casing (lower case)
    We check that the result is still equivalent (by lowering the result)
    We check that the path has been changed (expecting that it is the good case)

PS: The test on @"C:\pagefile.sys" has been removed because I discovered that the call to the native api are not working with files at the root of the drive 👿 😭