fix: Unable to intstall new version
There are few issues with the install script related to configuring
the telemetry settings.

1. On W7 there is only PowerShell 2.0 (that is what ships with .NET 2.0).
On the later versions of OS there are different versions of PowerShell and
thus requiring a specific version would lead to failures.

2. A problem related to the logic responsible for self-removal

  * First `$MyINvocation.InvocationName` returned different and unexpected
    values when a script is run not from a command line but invoked as a
    script-block. In the latter case, `$MyINvocation.InvocationName` is
    set to `&`.

  * Secondly the installer somehow would fail to remove the script
    (permissions?)

The self-removal logic is removed.

Also bumping the version to 3.2.1 to ensure the new fix flows through
to all distribution channels correctly.

Resolves #7120