fix: missing method ResourceManager.GitPluginBase..ctor
Fix a refactor regression - the translation app requires a parameterless
constructor.

Also add safe guards to the type iteration routine that ignore any failures.
Locally a lot of contributors would have custom plugins installed (e.g.
GitExtensions.PluginManager) that does not belong to the main codebase
and, thus, does not need to be translated.
The translation app would previously fail on these external types, and
fail to udpate the translations as a result.

Closes #7286