fix: FormResolveConflicts incorrect row selection
If a user selected a different row from the top, upon rescan/refresh of
the dialog there would be 2 rows selected - the top and the one that
the user wanted.
This would lead to a situation where an incorrect file would be shown,
or worse, no file would be shown due to multiple selection.

Fix the selection by resetting the initial selection after databind and
selecting the user's choice.
Also if the user selected the last row, and a conflict was resolved,
retain the selection and select the new last row upon refresh.