CI: Only look for newly finished builds if previously running build found
Makes GitExtensions more CI server friendly by doing less api calls:
=> prevent 1 (most of the time) useless api call / minutes

These calls are most of the times useless because finished builds
that will be found corresponds to commits not in the revision grid.

The only case where it is useful is when there was running builds the last time the revision grid was refreshed
 and so, to display the state of these builds once finished.

That's why this commit makes that the calls to find newly finished builds to be made only when running builds were previously found.

Contribute to fix too intensive call of CI api. Follow up of #7477 (For issue https://github.com/gitextensions/gitextensions/pull/7452#issuecomment-559084273 )