perf: Improve speed of directory change in a consoles
Fix #7595

Sending a lot of backspace in a console is slow
(especially with "cmd" but also with "bash" or "powershell").

Now, use a better startegy to send a char to go to the begin of the console line,
comment the line then evaluate it before launching the "cd command.

Which is quicker...