AzureDevOps CI: Better choose the running build to display
when there is more than one build to display for a commit.

Before:
Multiple build was return
and due to the condition in `BuildServerWatcher.cs`/`OnBuildInfoUpdate()`,
the last started build is displayed.

Now:
Return only one build by commit to better select the one displayed
which is now the first started build (or at default one not still started).