SidePanel Submodules: Status for top module
If there are changes in any submodules (commit or dirty) or changes in the topmodule itself, this status should be seen in the top most icon.
(The top module cannot be ahead/behind only dirty but will be so for any changes)

There are some limitations for how and when the icon is updated.
See #7779 for details

Refactor GitUI/BranchTreePanel/RepoObjectsTree.Nodes.Submodules.cs
 * Remove UseFolderTree, unused
 * Refactor: Simplify Submodule GetImage