Destroy build cache when moving to another repository
By settings a sort of "watchdog" flag.

If we detect that we still are in the same repository
and so that we still need the cache, we set the flag.
That prevent the cache to be disposed when the previous
instance of a `IBuildServerAdapter` is disposed in
`BuildServerWatcher.LaunchBuildServerInfoFetchOperationAsync()`.

If we don't succeed at setting the flag on time (because
we switched of repository and so we don't use the `AzureDevOpsAdapter`),
then the cache is disposed :fireworks: