DiffTool for large files
A few forms uses the FileViewer OpenWithDifftool Action to start the difftool.
The result if the form internal handler or the FileViewer implementation is used
should be the same, starting difftool from GitModule

This is the case for the HotKey (F3) handling in FormCommit and both
HotKey and context menu for RevisionFileTreeControl

The difftool must be set also when displaying the "empty" message.