Align context menu order for FileStatusList
Align the menu items in the menus for RevisionDiffControl,
RevisionFileHistory and FormCommit. This is mostly aligning the
order, but a few obsolete items has been disabled and removed.
Also some handling to change the default item for submodules to be first and
to be bold when default. (Added a TODO for the handling of this item,
and why it is not implemented for File History).

FormCommit: Do not set OpenSubmodule as bold
as it is not the default, for Staged file status

FormCommit: Remove unstaged files separate openSubmodule menu item
Use the item in FileStatusList as in other forms

Remove submodule summary menu item
More information in the diff viewer already

Submodules: Put menu item first for 'Open with Git Extensions"

Share text for "Open with Git Extensions"

RevDiff: WorkTree: Move submodule menu items to top
Aligns with FormCommit
Set item order in priority with Update submodules first and Commit last

FormCommit: Move History to align with other forms

RevFile: Move History to align with other forms

RevFile: Align order OpenFolder menu items

RevFile: Move Archive from open folder to blame

Move SaveAs menu item to OpenWith

FormCommit: Move OpenDifftool to top of context menus
(below Open with Git Extensions for submodules)

FormCommit: Move InteractiveAdd close to ResetParts

FormCommit: Hide editing of files that do not exist
For instance for submodules

FormCommit: Move Edit before OpenWith

FormCommit: Align OpenFolder with other forms