Redesign script configuration settings
* Update the layout to use a listview for listing scripts and a
propertygrid for editing script details

* One of a key change is that script modifications are no longer
"destructive" until the settings are accepted via [OK] or [Apply] buttons.
The race condition where a user has multiple instances of the app open
and modifies scripts in several instances at once still remains, and it
completely outside the scope of this PR.
The last instance of the app to accept changes to scripts will overwrite
all other changes may have been made in other instances.

* The UI is built under an assumption that a majority of our users
won't have more than a dozen of scripts.
Absent of any real data, this is based on personal observations.