Restore broken AutoCRLF tests
Resolves #8000

The tests got broken for few reasons:

1. Earlier versions of ApprovalTests considered differences in EOL, and
these tests deal with precisely this.
However recently we have updated to the newer version of ApprovalTests
that no longer considered EOL differences as failures.

2. During migration to Arcade SDK (#7649) few approved files had their
EOL accidentally converted to CRLF, that was unnoticed due to the reasons
explained above.

CRLF conversion is moved out of `FileViewer` because it has no real
connection to it.
The tests refactored and verified for correctness.