fix: New remotes not saved, if exceed limit
If a user adds a new remote which index, when saved, exceeds the number
of maximum recent remotes - the remote would be added to the end of the
list of remotes, and trimmed off when the list of recent remotes is saved.

Fix the issue by adding the new remote to the top of the list, so it gets
saved.