Reorder context menu
Move "Open with diff tool" context menu out from the top position
in FormCommit.

In the File Tree and Diff tabs a user would typically deal with existing
commits (i.e. snapshots), and "compare" function is likely to be one of
the most used (i.e. compare to the working directory).

In the Commit form the user is dealing with the working directory changes,
and thus staging and reseting is likely be the primary function, then
compare (which is rendered in the diff panel of the form).

Resolves #8106