Display 'Plugin Manager" next to "Plugin settings" in the menu
to make it more visible

Sort plugins in descending alphabetic order to add them **always**
at the top so that they end up appearing in alphabetic order
(without having to handle an index to where insert them).

Do an exception for the plugin manager which is add just above the plugin settings item.