Clean up of status reporting of failed operations
Undo plumbing and infra introduced in 1572cbd7da7af53678f05b8cd4dcfcdf9e8e46d3.
With that we have one less dependency on WPF stack.

There are several scenarios that involve showing `FormStatus` or its descendants.
1. Show a progress as a modal dialog (this is standard way).
2. Do not show the progress dialog, and only show it if an operation that
was running failed (e.g. staging files in `FormCommit`).

The approach to showing results of failed operations appears to be unnecessay
complex, that IIUC led to issues while attempting to show a dialog that was
already considered opened but invisible.
And the WPF's DispatcherFrameModalController was used to deal with this.
This also necessitated defining delegates, and required more plumbing to
pass those around, validation, etc.

However it is much simpler *not* to show the dialog *until after* the operation
has failed.
The callsites are also much easier to read and understand.