Hide filtered out refs
Indicate branches and tags in the left panel that are currently filtered
out and not visible in the revision grid.

* The hidden refs are greyed out and their icon is changed to "closed eye".
* Each node in the tree now has a new property `IsVisible` that can be
used as a conditional flag when implementing necessary logic.
* Apply the same style to branches in the Recent branches toolstrip dropdown.

Detection of whether a given ref is visible or not is done by checking
the revision grid, which is currently the source of truth.
Whenever the revision grid is refreshed, it raises an event that causes
the left panel to update its state, and show or hide nodes.