fix: Resolve script "c"-variables
Whenever a script, run outside the revision grid, needed to resolve
a "c"-variable, it would  and Original implementation would retrieve
the current checkout information, that didn't have enough information
available to process a requested variable.

Replace the call with `GetRevision` call, that is made by the revision grid,
and which has all information about a revision (except for the full body).
With that consolidate the revision retrieval logic.

Expose `GetRevision` method at the `IGitModule` level, and move
`GitRevision`type to GitUIPluginInterfaces, so it is universally available.

Fixes #8486