Fix JumpList being created too soon
Currently, at least on some systems, there is no process-level main
window yet when FormBrowse is loaded, so the creation of the JumpList
fails ("A valid active Window is needed").

So we wait for the OnActivated(), where this seems to be always OK.
We also need to keep track of any call to AddToRecent() it it happened
before actual JumpList creation.

Fixes #8421