Don't refresh revision grid invoking toolstrip actions
Appears to be a regression from #8485 which caused the revision grid
to be refreshed whenever an action was raised via the toolstrip, e.g.
invoking fetch would refresh the grid before the fetch operation was
performed, as well as after.

Somehow the toolstrip would pass focus to all its items, that included
the Branch Filter which in turn would cause the refresh.
Update the logic to force the grid refresh only when the filter is being
changed and applied.

What's worse, multiple refreshes could overlap each other due to peculiar
behaviour of the toolstrip, which keeps on passing focus back to its items,
and with that invoke `FilterBranchHelper.toolStripBranches_Leave()`.
There was also another strange behaviour that could lead to a blank revision
grid - if a user typed a filter but hasn't applied it, and then invoked
another action via the toolstrip (e.g. fetch) - the revision grid would
refresh first after the fetch, and then another refresh operation would
be in flight that attempted to apply the pending filter.

Remove the method and in essence change the user behaviour - now users
have to hit ENTER to apply a filter rather than leave the filter textbox.
And move the latch in `ForceRefreshRevisions()` a little higher, to
reduce risks of race refreshes.

Also do not refresh the grid when fetching the status from the superproject
if the status hasn't changed.