Break the toolstrip (#8572)
* Break the main toolbar into two - general operations and filters. 
This is arbitrary, filters were the easiest to break out. The main point of the PR was to attempt the split and see how well it works. This then maybe used as a foundation for further refactoring of the toolstrip area, such as moving user scripts to own toolstrip, view-related, repo-related... Possibilities are endless.
* Save/restore positions of the toolbars
* Allow show/hide of the toolbars via a contextmenu and the main menu (under "View")