Do not  throw if Git is upgraded  while GE is started
Ignore exceptions for GitStatusMonitor and Background fetch plugin,
as well as "Browse activation" (selecting the window).
This is the most common background use of Git.
The primary usage is to allow Git upgrades while GE is active.

Active use of GE will raise exceptions, not much can be done when Git is not available.