Better exception presentation
Trap and route as many exceptions as possible to the central handler.

For exceptions coming from the main thread it is possible to ignore those
and continue the app's execution, but for unhandled exceptions coming
from background threads the app will terminate.

The main focus of this work is to intercept exceptions originated from
`Executable` (i.e. executing git commands, etc.), and user scripts.
However it is not limited in this scope.

Contributes to #7795
Co-authored-by: Michael Seibt <mstv@users.noreply.github.com>