AzureDevOps CI: Cache finished builds results (to prevent a costly call at every grid revision refresh)
and reduce the scope of the first api call that retrieve all the finished builds
* Keep in cache only the last finished build per commit
* Add a new method called when changing from repository

BREAKING CHANGE: Change the `IBuildServerAdapter` interface by adding a new method
and so will break the external BuildIntegration plugins
(like https://github.com/martin211/GitExtensions.GitLabPipelinesPlugin ).
They will require a new release (but maybe it's the good moment to do so
because they will anyway have to do a new release built with .net5

+ BuildServerWatcher: introduce a lock to fix some builds result not displayed

because early exit in `OnBuildInfoUpdate()`
due to `_buildStatusCancellationToken == null`.

Indeed, if the build results arrived immediatly,
`_buildStatusCancellationToken` is still not initialized
(because it will be when object is created)
when the first results will be processed.

The solution is to `lock(){}` so that it insure
that `_buildStatusCancellationToken` is set
when processing the first build result...

+ Remove forgotten file TfsAdapter.cs (Follow up of 8190d0c0570af95ffbf9e168072d1b169eb99618)