fix some builds result not displayed
because early exit in `OnBuildInfoUpdate()`
due to `_buildStatusCancellationToken == null`.

Indeed, if the build results arrived immediatly,
`_buildStatusCancellationToken` is still not initialized
(because it will be when object is created)
when the first results will be processed.

The solution is to `lock(){}` so that it insure
that `_buildStatusCancellationToken` is set
when processing the first build result...