Variety of accessibility-related fixes for config dialog
1. Update app.config to pick up some .NET Framework accessibility fixes up to 4.8,
depending on what version of .NET Framework is available on the device at runtime.

2. Add unique accessible names for the Browse and Suggest buttons to avoid duplicate
names being exposed programmatically on sibling buttons. (If the UI is ever localized,
the accessible names must be localized too.)

3. Change the order in which controls are added to their containers, in order to have
the programmatic order (as exposed through the UI Automation API) of the controls match
their logical order.

4. Remove the TabIndex property being set on the controls, because that interferes with
WinForms setting accessible names on TextBox and ComboBox controls based on the text from
the Labels that programmatically precedes them. The default tab order is now based on the
order in which the controls are added to their containers.

(cherry picked from commit bbddb2de2f9933b4531bb97eea48b967ca453637)