fix: Select diff when there are staged and unstaged files
If there are both stages and unstaged files the commit dialog fails
to select the first unstaged file, thus showing no diff.
If there are only unstaged or staged files everything works as expected.

When the form opens it expects "Commit" button to have focus (i.e. be
an active control), however it is not happening, and the active control
remains null.

Treat "no active control" the same way as "active control == Commit"
to force the selection.