Correct ownership of progress dialogs
In multi monitor scenarios, where the main form is on a 2ndary monitor,
checking out a branch without showing the checkout dialog could lead to
situations where the checkout progress dialog would be shown on the
primary screen.

This was caused by setting the progress dialog's owner to `null`.

Ensure the correct owner is passed to the progress dialog in the checkout
branch sequence.

Also correct various callsites in the revision grid and the left panel to
pass the form as the owner instead of respective controls.