Introduce new interface for named git refs
This comes ahead of work to null-annotate these APIs. Not all implementations will populate this property, so in the old factoring it should be marked as nullable. However the Name property is used in many places where it is assumed non-null.

This factoring splits out those usages behind a new interface, making it easier to correctly annotate these APIs.