AppVeyor: Handle v2 tokens
v2 tokens require that the account is prepended to the request..
This apparently is only required for quering repos for an account,
not normal info.

* Only query for project names for a configured account if no
account/repo at all is specified
Illegal repos will not report any builds anyway.
This allows a AppVeyor username/token at the same time as multiple
repos are used. For instance can gitextensions/gitextensions builds
be seen at the same time as user private builds

* Allow configuring of  account/repo also if accountName is set

* Less aggressive requery of running AppVeyor jobs

* Refactor
  * remove static variables/functions
  * Handle projectId/name as account/repo consistently
  * Use local functions (and inline code)
  * Remove internal data class Project (the module global is no longer needed)