Fix ownership of FormUpdates
`FormBrowse` is the main form, and hence it doesn't have `Owner` set.
This leads to `FormUpdates` being created as an orphant modal form,
that can be hidden behind `FormBrowse` and making the app look
unresponsive.

Correct the ownership and, thus, force the new form to show above
the main form.