Submodule: Recreate tree only at structure changes
This applies both for the sidepanel and browse toolbar menu.
The submodule provider stsus updates always contains the full trees,
but as the provider notifies when a new list is being created,
sidepanel/toolbar-menu can know when a new structure must be created.

* Browse submodule menu: Refactor

Remove obsolete submodule detail handling
SubmoduleStatusProvider calculates the required info.