Rename translation strings to avoid conflict
So long as we target .NET Framework, we do not have annotated versions of string.IsNullOrEmpty and friends. We've worked around this by introducing a `Strings` class that has annotated versions of these, but that class had a naming conflict with existing translation string resources.

This commit renames the translation strings to avoid the conflict, which simplifies several usages of the `Strings` type.