Update Clone dialog's handling of clipboard text
This change adds a new FormClone.TryExtractUrl method
to split a string by spaces and extract the first string for which
Uri.IsWellFormedUriString returns true.

A new integration test is added for FormClone to validate the
behavior of the new method.

This change updates the Clone dialog to use the new method
for extracting URLs from clipboard contents.

Calls to PathUtil.IsUrl are replaced by direct calls to
Uri.IsWellFormedUriString.