Report bugs in a separate process
- If an unhandled exception is throw on the main thread, show the error
within the app process. The error may or may not be fatal.
This code path is for the most part unchanged from the existing
behaviour.

- If an unhandled exception is thrown from a background thread it
will crash the app. Serialize and encode the error information as base64
string, and launch the bug reporter app before terminating the app.

- The bug report exception handling is updated to use SerializableException
type to accommodate both -in-app and out-of-proc exception reporting.