Reduce window flicker and loss of focus
1. Force child window ownership to FormBrowse

2. Show the rebase dialog *before* starting any operations. This is what
appears to be the main cause of the loss of focus. The dialog isn't yet
shown when we invoke the progress dialog, which is then gets closed
but the window that spawned it is not visible. MS Windows gets confused
and pushes another window to the foreground, thus causing the app to
lose focus.