GitRevision: Remove Name property
Name was used in FormFileHistory for a specific file only,
not fitting the general use in RevisionReader.
Instead just parse this data in FormFileHistory where it is used:

* Displaying a file.
  - Set title (if it differs).
  - Check if the file exists on the filesystem (affects operations on submodules and line patching).

* Diff a file, select the old filename in an older revision.

* Use old filename in SaveAs dialog.

The parsing in StreamExtensions was incorrect for git-log -z --name-only,
only returning the first filename and possibly other filenames as separate
"commits" and an empty last marker.
The extra filenames and the empty end were giving parse errors
that had to be ignored.

The stream could be splitting on double null terminated records,
in the used scenarios, but that is not reliable for commits without
filnames (like stash).

When skipping the generic case (and variable length commit message),
this is much simpler.

This allowed for removing EndOfBody handling in RevisionReader,
simplifying the handling.