RevisionReader: Parse raw commit body
* Parse raw body %B instead of combining %s%n%n%b to avoid extra processing
in both Git and GE.
Simplify the handling.

* Do not allocate body for older commits
Previously a temporary variable was allocated also if body was
not stored in the commit message (to save memory, for older commits).

* Only store commit body if multiline message
If the body is the same as the subject, do not store the message
This also avoids loading the commit body for older commits
if the commit is selected

* Align usage of commit body retrieval for other than RevisionReader
  * "Previous commit messages" used %s%b instead of %B
  * Align parsing of raw body and subject

* ObjectId Inline local parsing functions

* Use ReadOnlySpan<char> in RevisionReader
Reduce string allocations

* Use HighPerformance StringBuffer