Do not blame folders
It is possible to blame Git Tree objects (folders) from File tree,
then the Diff/View/Blame could possibly only be shown for one "random"
file in each commit (or none as git-log --follow do not report all
commits).
To decrease confusion only one of the files are shown.

For artificial commits, this means that no tabs are displayed.

* If a file object was blamed, diff the same file for Artificial commits
as for HEAD.