Do not blame folders
It is possible to blame Git Tree objects (folders) from File tree
GE shows the history for only one of the files.
If no files are seen for a certain commit, GE uses the "FileHistory"
path for blame, which is not possible.

In this case hide the Diff, View and Blame tabs.
For artificial commits, this means that no tabs are displayed.