Diff to actual parent in filtered grids
If the grid is filtered, the commit parents are rewritten to show the
displayed parent commit.
The diff in RevDiff should show the diff for the actual commit rather than
the diff to the previous node.

Some refactoring and related updates:

* Diff: Avoid calculate HEAD, available in RevGrid

* ShowDiffForAllParents menu item used partially initiated info
The functionality was correct though.

* Add method to get GitRevisions with actual parents if revGrid
has rewritten the parents in filtered views.

* Use information from revGrid if available before requesting from Git