Support WSL Git
For Git repos stored in \\wsl$ directories, execute the Git executable in WSL
where possible to improve performance. WSL Git commands are several
times faster when executed by the WSL Git executable instead of the
Git Extensions Windows (native) application.

Using the WSL Git executable requires that paths provided to WSL Git
are adjusted from Git Extensions Windows (native/internal) format to WSL format,
like \\wsl$\Ubunto\repo to /repo and c:\repo to /mnt/c/repo.
The paths internal to Git Extensions are always in Windows format.
Therefore, paths in both inputs and outputs for Git commands must be translated.

The Git Extensions Windows (native) Git executable is still used for the
following:
* All handling and settings related to Git in Settings. This includes
display of Git version as well.
However, if the Git version is too old to be supported, the GitVersion
check added in #9680 will report that the version is too old. (This
change is in a separate commit.)
* Custom merge implementation in FormResolveConflicts.
* Custom difftool/mergetool, the tools are only extracted in Windows.
Default git-difftool is using the Git executable, which could be separate
for WSL Git.
* Some built-in plugins etc, ScriptRunner and FindLargeFiles always use
 Windows Git

A special note for worktrees:
A relative path is used at creation.
This will work with both Windows Git and wsl Git.
Note that Git will use a full "native" path internally regardless,
that has to be edited if the worktree is to be used in both
Win/WSL, but the worktree is usable with the Git executable creating the path.