SHIFT+DEL = quick delete branch
Currently it is a hassle to remove a number of obsolete branches.
There is so much clicking and confirming necessary.
With this PR you simply press SHIFT+DEL it will:
  - choose the first available branch for the current commit for deletion
  - enable the "delete remote branch" checkbox
  - suppress (quietly confirm) confirmation dialogs

I have not signed the "The Developer Certificate of Origin". But:
Hereby, I declare that I claim no rights on this commit.
Instead I donate my changes free of charge and free of demands of any kind.
Any developer may use, rebase, or merge these changes.