Move the init logic from OnLoad to InternalInitialize
If the app is opened with `browse <repo>` we would execute a number of
instructions that we wouldn't execute if the app was opened in the
dashboard-mode. Those instructions included initialising the "Push" button,
the revision grid.

Move these instructions into `InternalInitialize` method, that gets
executed whenever a dashboard is shown or a repo is loaded. This
way we unify the flow, and contain repo-init instructions in one place.

At the same time restore the splitter positions when we load the form
irrespective of the current view.

Addendum to #9106