Invoke UICommandsChanged event when loading the form
`FormBrowse` subscribed to `UICommandsChanged` events, however
this event was not fired when the form initially loaded, only after a
repository is switched (either set from the dashboard or selecting a
different repository).

This was because `GitModuleForm` would assign the passed in `GitUICommands`
to a field instead of the property, which raises the event.

Correct this, assign the passed in commands to the property, and raise the event.
React to the the event in `FormBrowse` via the `OnUICommandsChanged` override.