Fix loading user plugins.
Before this change the user plugins path was set in the PluginRegistry, which was initialized in the FormBrowse and the FormCommit. The MEF catalog is initialized earlier in the Program.RunApplication and so user plugins are never loaded.

This change moves initialization of user plugins path to the MEF initialization.